+++ lib/src/lib/enums.dart
@@ -1,3 +1,3 @@
 enum TapeType { unknown, tap, tzx }
 
-enum AudioFilterType { none, heuristic, bassBoost, tapir }
+enum AudioFilterType { none, heuristic, bassBoost, tapir, sine }